mesh: platform-neutral stream pair + widen the stream ABI (Windows groundwork)
Groundwork for running the mesh backends on Windows. No behaviour change on
Linux/macOS — the Unix data plane is byte-for-byte the same path as before.

New shim/meshpair module. The creation of the local stream pair (whose two ends
join the C++ pump to the Go bridge) moves out of the ts/nb shims into a module
that has no cgo, no mesh libraries and no DuckDB. That matters because it can
then be exercised with plain `go test` on every OS in seconds — the Windows
loopback pair's accept-matching and half-close semantics are the trickiest new
code in the port, and this is the only test that covers them directly.

  pair_unix.go     verbatim lift of the proven AF_UNIX socketpair path
  pair_windows.go  loopback TCP pair (no socketpair() exists on Windows)

Windows specifics that are easy to get wrong, so they are in from the start:
  - C's end is a raw windows.Socket, never a net.Conn. net.FileConn does not
    accept a Windows SOCKET at all, and any socket the net package owns would be
    closed under C when its conn is closed or collected.
  - SO_EXCLUSIVEADDRUSE on the listener: on Windows another local process can
    otherwise bind the same loopback port and steal connections. No Unix analogue.
  - TCP_NODELAY on both ends: loopback TCP applies Nagle + delayed ACK, which
    would add ~40ms stalls to request/response traffic that AF_UNIX never had.
  - the accept is gated on the full TCP 4-tuple, so a local process racing us to
    the ephemeral port is detected and dropped rather than handed to C.

ABI: `int *fd_out` becomes `mesh_stream *` (uintptr_t) end to end — a Win64
SOCKET is UINT_PTR and does not fit an int. Verified the cgo-generated headers
still match the hand-written contract in shim/mesh_shim.h.

Two real bugs found on the way:
  - the embedded blob length was declared `unsigned long` against an 8-byte
    .quad; that is 4 bytes under MSVC's LLP64, i.e. a silently truncated length.
    Now uint64_t.
  - dlerror() was called twice in one expression, so every load failure built a
    std::string from the second (NULL) call — UB. Read once.

Verified on Linux with mesh=both: core_tests, the SQL suite against the docker
sshd, the lazy-load proof (SSH-only sessions still map no Go), the single-mesh
latch in both orders, and — the ones that actually exercise the changed dial
path — real WireGuard payloads over both the Headscale/Tailscale and the
self-hosted NetBird data planes.

+++ src/mesh_forwarder.cpp
@@ -97,9 +97,10 @@ void MeshForwarder::AcceptLoop() {
             close(client);
             break;
         }
+        // MeshStream is an fd here; the Winsock port replaces this with socket_t.
         int mesh_fd = -1;
         try {
-            mesh_fd = backend_->Dial(attrs_.remote_host, attrs_.remote_port);
+            mesh_fd = static_cast<int>(backend_->Dial(attrs_.remote_host, attrs_.remote_port));
         } catch (const std::exception &e) {
             attrs_.error_message = e.what();
             close(client);
